+++ src/azure-cli-core/azure/cli/core/_session.py
@@ -4,6 +4,7 @@
 # --------------------------------------------------------------------------------------------
 
 import json
+import logging
 import os
 import time
 
@@ -46,9 +47,17 @@ class Session(collections.MutableMapping):
                     self.save()
             with codecs_open(self.filename, 'r', encoding=self._encoding) as f:
                 self.data = json.load(f)
-        except (OSError, IOError, t_JSONDecodeError):
-            get_logger(__name__).warning("Fail to load or parse file %s. It is overridden by default settings.",
-                                         self.filename)
+        except (OSError, IOError, t_JSONDecodeError) as load_exception:
+            # OSError / IOError should imply file not found issues which are expected on fresh runs (e.g. on build
+            # agents or new systems). A parse error indicates invalid/bad data in the file. We do not wish to warn
+            # on missing files since we expect that, but do if the data isn't parsing as expected.
+            log_level = logging.INFO
+            if isinstance(load_exception, t_JSONDecodeError):
+                log_level = logging.WARNING
+
+            get_logger(__name__).log(log_level,
+                                     "Failed to load or parse file %s. It will be overridden by default settings.",
+                                     self.filename)
             self.save()
 
     def save(self):