+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/_params.py
@@ -42,7 +42,6 @@ from ._constants import (
     MANAGED_REGISTRY_SKU,
 )
 from ._validators import (
-    validate_registry_name,
     validate_headers,
     validate_build_arg,
     validate_secret_build_arg
@@ -78,6 +77,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-statements
         c.argument('build_arg', help="Build argument in 'name[=value]' format.", action='append', validator=validate_build_arg)
         c.argument('secret_build_arg', help="Secret build argument in 'name[=value]' format.", action='append', validator=validate_secret_build_arg)
         c.argument('no_logs', help="Do not show logs after successfully queuing the build.", action='store_true')
+        c.argument('no_format', help="Indicates whether the logs should be displayed in raw format", action='store_true')
         c.argument('os_type', options_list=['--os'], help='The operating system type required for the build.', arg_type=get_enum_type(OsType))
 
     with self.argument_context('acr import') as c:
@@ -110,7 +110,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-statements
         c.argument('image', arg_type=image_by_tag_type)
 
     with self.argument_context('acr create') as c:
-        c.argument('registry_name', completer=None, validator=validate_registry_name)
+        c.argument('registry_name', completer=None)
         c.argument('deployment_name', arg_type=deployment_name_type, validator=None)
         c.argument('location', arg_type=get_location_type(self.cli_ctx), validator=get_default_location_from_resource_group)