+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/tests/latest/test_webapp_commands_thru_mock.py
@@ -55,7 +55,7 @@ class TestWebappMocked(unittest.TestCase):
     def test_set_source_control_token(self, client_factory_mock):
         client_factory_mock.return_value = self.client
         self.client._client = mock.MagicMock()
-        sc = SourceControl('not-really-needed', source_control_name='GitHub', token='veryNiceToken')
+        sc = SourceControl(name='not-really-needed', source_control_name='GitHub', token='veryNiceToken')
         self.client._client.send.return_value = FakedResponse(200)
         self.client._deserialize = mock.MagicMock()
         self.client._deserialize.return_value = sc
@@ -70,13 +70,13 @@ class TestWebappMocked(unittest.TestCase):
     def test_set_domain_name(self, client_factory_mock):
         client_factory_mock.return_value = self.client
         # set up the return value for getting a webapp
-        webapp = Site('westus')
+        webapp = Site(location='westus')
         webapp.name = 'veryNiceWebApp'
         self.client.web_apps.get = lambda _, _1: webapp
 
         # set up the result value of putting a domain name
         domain = 'veryNiceDomain'
-        binding = HostNameBinding(webapp.location,
+        binding = HostNameBinding(location=webapp.location,
                                   domain_id=domain,
                                   custom_host_name_dns_record_type='A',
                                   host_name_type='Managed')
@@ -96,13 +96,13 @@ class TestWebappMocked(unittest.TestCase):
         client_factory_mock.return_value = client
         cmd_mock = mock.MagicMock()
         # set up the web inside a ASE, with an ip based ssl binding
-        host_env = HostingEnvironmentProfile('id11')
+        host_env = HostingEnvironmentProfile(id='id11')
         host_env.name = 'ase1'
         host_env.resource_group = 'myRg'
 
         host_ssl_state = HostNameSslState(ssl_state=SslState.ip_based_enabled, virtual_ip='1.2.3.4')
-        client.web_apps.get.return_value = Site('antarctica', hosting_environment_profile=host_env,
-                                                host_name_ssl_states=[host_ssl_state])
+        client.web_apps.get.return_value = Site(name='antarctica', hosting_environment_profile=host_env,
+                                                host_name_ssl_states=[host_ssl_state], location='westus')
         client.app_service_environments.list_vips.return_value = AddressResponse()
 
         # action
@@ -113,8 +113,8 @@ class TestWebappMocked(unittest.TestCase):
 
         # tweak to have no ip based ssl binding, but it is in an internal load balancer
         host_ssl_state2 = HostNameSslState(ssl_state=SslState.sni_enabled)
-        client.web_apps.get.return_value = Site('antarctica', hosting_environment_profile=host_env,
-                                                host_name_ssl_states=[host_ssl_state2])
+        client.web_apps.get.return_value = Site(name='antarctica', hosting_environment_profile=host_env,
+                                                host_name_ssl_states=[host_ssl_state2], location='westus')
         client.app_service_environments.list_vips.return_value = AddressResponse(internal_ip_address='4.3.2.1')
 
         # action
@@ -125,8 +125,8 @@ class TestWebappMocked(unittest.TestCase):
 
         # tweak to have no ip based ssl binding, and not in internal load balancer
         host_ssl_state2 = HostNameSslState(ssl_state=SslState.sni_enabled)
-        client.web_apps.get.return_value = Site('antarctica', hosting_environment_profile=host_env,
-                                                host_name_ssl_states=[host_ssl_state2])
+        client.web_apps.get.return_value = Site(name='antarctica', hosting_environment_profile=host_env,
+                                                host_name_ssl_states=[host_ssl_state2], location='westus')
         client.app_service_environments.list_vips.return_value = AddressResponse(service_ip_address='1.1.1.1')
 
         # action
@@ -142,7 +142,7 @@ class TestWebappMocked(unittest.TestCase):
         client_factory_mock.return_value = client
 
         # set up the web inside a ASE, with an ip based ssl binding
-        site = Site('antarctica')
+        site = Site(name='antarctica', location='westus')
         site.default_host_name = 'myweb.com'
         client.web_apps.get.return_value = site
 
@@ -174,7 +174,7 @@ class TestWebappMocked(unittest.TestCase):
         # Mock the client and set the location
         client = mock.Mock()
         client_factory_mock.return_value = client
-        site = Site('antarctica')
+        site = Site(name='antarctica', location='westus')
         site.default_host_name = 'myweb.com'
         client.web_apps.get.return_value = site
 
@@ -195,7 +195,7 @@ class TestWebappMocked(unittest.TestCase):
 
     @mock.patch('azure.cli.command_modules.appservice.custom._generic_site_operation', autospec=True)
     def test_update_site_config(self, site_op_mock):
-        site_config = SiteConfig('antarctica')
+        site_config = SiteConfig(name='antarctica')
         site_op_mock.side_effect = [site_config, None]
         cmd = mock.MagicMock()
         # action
@@ -220,10 +220,10 @@ class TestWebappMocked(unittest.TestCase):
     @mock.patch('azure.cli.command_modules.appservice.custom.get_streaming_log', autospec=True)
     @mock.patch('azure.cli.command_modules.appservice.custom.open_page_in_browser', autospec=True)
     def test_browse_with_trace(self, webbrowser_mock, log_mock, site_op_mock):
-        site = Site('antarctica')
+        site = Site(location='westus', name='antarctica')
         site.default_host_name = 'haha.com'
         site.enabled_host_names = [site.default_host_name]
-        site.host_name_ssl_states = [HostNameSslState('does not matter',
+        site.host_name_ssl_states = [HostNameSslState(name='does not matter',
                                                       ssl_state=SslState.ip_based_enabled)]
 
         site_op_mock.return_value = site