+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/_help.py
@@ -19,7 +19,7 @@ helps['ad sp create-for-rbac'] = """
         - name: --cert
           short-summary: Certificate to use for credentials.
           long-summary: When used with `--keyvault,` indicates the name of the cert to use or create.
-            Otherwise, supply a PEM or DER formatted public certificate string. Use `@{file}` to
+            Otherwise, supply a PEM or DER formatted public certificate string. Use `@{path}` to
             load from a file. Do not include private key info.
         - name: --create-cert
           short-summary: Create a self-signed certificate to use for the credential.
@@ -86,7 +86,7 @@ helps['ad sp credential reset'] = """
         - name: --cert
           short-summary: Certificate to use for credentials.
           long-summary: When using `--keyvault,` indicates the name of the cert to use or create.
-            Otherwise, supply a PEM or DER formatted public certificate string. Use `@{file}` to
+            Otherwise, supply a PEM or DER formatted public certificate string. Use `@{path}` to
             load from a file. Do not include private key info.
         - name: --create-cert
           short-summary: Create a self-signed certificate to use for the credential.
@@ -154,6 +154,10 @@ helps['ad app update'] = """
 helps['ad user list'] = """
     type: command
     short-summary: List Azure Active Directory users.
+"""
+helps['ad user get-member-groups'] = """
+    type: command
+    short-summary: Get groups of which the user is a member
 """
 helps['role'] = """
     type: group