+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/tests/latest/test_storage_scenarios_non_ascii_chars.py
@@ -6,7 +6,7 @@
 
 import os
 from azure.cli.testsdk import LiveScenarioTest, ResourceGroupPreparer, StorageAccountPreparer, JMESPathCheck
-from .storage_test_util import StorageScenarioMixin
+from ..storage_test_util import StorageScenarioMixin
 
 
 class StorageCopyTests(StorageScenarioMixin, LiveScenarioTest):